Backend/requirement.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
python-dotenv
Flask-JWT-Extended
Flask-Bcrypt
Flask-Cors